requirements.txt
anyio==4.8.0
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mutagen==1.47.0
numpy==2.2.3
opencv-python==4.11.0.86
python-telegram-bot==22.0
pytubefix==9.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.7.0
mutagen==1.47.0
nodejs-wheel-binaries==22.20.0
numpy==2.2.3
opencv-python==4.11.0.86
propcache==0.4.1
python-telegram-bot==22.0
pytubefix==10.1.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
yarl==1.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.7.0
mutagen==1.47.0
nodejs-wheel-binaries==22.20.0
numpy==2.2.3
opencv-python==4.11.0.86
propcache==0.4.1
python-dotenv==1.1.1
python-telegram-bot==22.0
pytubefix==10.1.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
yarl==1.22.0
